clean code certification component
